extract out all of the connection handling into the proxy module
